METRON-1741 Moved REPL port of the Profiler to its own project
