+++ metron-analytics/pom.xml
@@ -47,6 +47,7 @@
 		<module>metron-profiler-client</module>
 		<module>metron-profiler-common</module>
 		<module>metron-profiler-spark</module>
+        <module>metron-profiler-repl</module>
     </modules>
 	<dependencies>
 		<dependency>
